src/genie_project/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.2; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
mysqlclient==2.2.5; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
